move controlpanel to interactive, add docstrings
